# leaf-types cache for ml-matches