--- a/src/dump.c
+++ b/src/dump.c
@@ -2237,9 +2237,6 @@ STATIC_INLINE jl_value_t *verify_type(jl_value_t *v) JL_NOTSAFEPOINT
 }
 #endif
 
-jl_datatype_t *jl_lookup_cache_type_(jl_datatype_t *type);
-void jl_cache_type_(jl_datatype_t *type);
-
 static jl_datatype_t *jl_recache_type(jl_datatype_t *dt) JL_GC_DISABLED
 {
     jl_datatype_t *t; // the type after unique'ing