--- a/src/precompile.c
+++ b/src/precompile.c
@@ -314,12 +314,6 @@ static void jl_compile_all_defs(void)
     JL_GC_POP();
 }
 
-static int precompile_enq_all_cache__(jl_typemap_entry_t *l, void *closure)
-{
-    jl_array_ptr_1d_push((jl_array_t*)closure, (jl_value_t*)l->func.linfo);
-    return 1;
-}
-
 static int precompile_enq_specialization_(jl_method_instance_t *mi, void *closure)
 {
     assert(jl_is_method_instance(mi));
@@ -370,7 +364,6 @@ static int precompile_enq_all_specializations__(jl_typemap_entry_t *def, void *c
 static void precompile_enq_all_specializations_(jl_methtable_t *mt, void *env)
 {
     jl_typemap_visitor(mt->defs, precompile_enq_all_specializations__, env);
-    jl_typemap_visitor(mt->cache, precompile_enq_all_cache__, env);
 }
 
 void jl_compile_now(jl_method_instance_t *mi);