When we do a `methods` lookup and get back a single result, it's very easy for us to cache that information, since we essentially already have the data-structure for it (TypeMapEntry). This lets us shave off a bit of time on micro-benchmarks by putting this information into a hash table instead of a tree (which is still used to handle the general case):
```
julia> @btime methods(+, (Int, Int))
master:  3.322 μs (21 allocations: 912 bytes)
PR:      2.069 μs (18 allocations: 784 bytes)

julia> @btime Base._methods_by_ftype(Tuple{typeof(+), Int, Int}, -1, typemax(UInt64))
master:  1.496 μs (7 allocations: 464 bytes)
PR:      0.175 μs (4 allocations: 336 bytes)
```