@JeffBezanson 2020-06-19
Looks fine to me. No reason to block merging, but I have a couple thoughts.

I wonder if we can take this farther to memoize more of `abstract_call_gf_by_type` --- i.e. avoid the "double lookup" of ml_matches followed by typeinf_edge.